Split memory integration tests from workspace model
