+++ docs/DECISIONS.md
@@ -209,6 +209,7 @@
 - Tool-call execution routing lives in `WorkspaceToolCallExecutor`. Browser inspect, plan update, SSH Remote project dispatch, local `ToolRouter` fallback, and successful apply-patch review-diff follow-up share one tested routing boundary. `WorkspaceModel` owns context refresh, event recording, persistence, and top-bar status, not tool-name routing branches.
 - Project metadata loading lives in `WorkspaceProjectMetadataLoader`, not inline in `WorkspaceModel`. Local project refresh aggregates instructions, local environment actions, extension manifests, and project memories through one loader; SSH Remote refresh converts remote context into the same metadata contract with local-only actions/extensions cleared. `WorkspaceModel` owns selection, thread-context application, persistence, and top-bar refresh after metadata is loaded.
 - Pure project loader coverage lives in focused test files, not `WorkspaceModelTests`. Direct tests for `ProjectInstructionLoader`, `LocalEnvironmentActionLoader`, `ProjectExtensionManifestLoader`, and `MemoryNoteLoader` should stay in their matching test files; `WorkspaceModelTests` should prove workspace integration only.
+- Workspace memory integration coverage lives in `WorkspaceMemoryIntegrationTests`, not `WorkspaceModelTests`. Global/project memory loading, `/remember`, agent memory tool execution, credential-like memory rejection, memory deletion, and the memory-add command are one feature group and should stay tested together outside the model monolith.
 - Worktree request values, tool-call construction, and handoff transcript records live outside `WorkspaceModel`. `WorkspaceWorktreeRequests` owns the public create/remove request structs used by dialogs, SwiftUI, and desktop controllers. `WorkspaceWorktreeToolCallPlanner` owns canonical `host.git.worktree.create` and `host.git.worktree.remove` argument JSON, including branch/base trimming and force flags. `WorkspaceWorktreeOpenEngine` owns the pure local/SSH Remote `Worktree: ...` thread construction, notice payloads, and display labels. The workspace model owns only tool dispatch, local/SSH project registration, selection, persistence, and top-bar refresh after a worktree is created.
 - Workspace status copy and context labels live in `WorkspaceStatusTextBuilder`. `/status`, slash mode confirmations, top-bar subtitles, instruction labels, memory labels, and mode labels must use the same helper so transcript copy, SwiftUI surfaces, static HTML, and menu-bar surfaces do not drift. `WorkspaceModel` owns when to emit a status response; the builder owns what the status response says.
 - Top-bar status semantics live in `QuillCodeTopBarStatusPresentation`, not in SwiftUI or HTML renderers. Agent status tone, indicator visibility, accessibility text, and runtime issue tone are shared presentation values so native UI and static UI harness snapshots stay visually and semantically aligned.
